#27184031 Реализовать создание статуса в настройках представлений в workflows.

+++ LeokaEstetica.Platform.Models/Entities/Template/ProjectManagmentTaskStatusIntermediateTemplateEntity.cs
@@ -18,7 +18,7 @@ public class ProjectManagmentTaskStatusIntermediateTemplateEntity
     /// <summary>
     /// Id статуса.
     /// </summary>
-    public int StatusId { get; set; }
+    public long StatusId { get; set; }
 
     /// <summary>
     /// FK.
